backend/requirement.txt
asgiref==3.9.0
Django==5.2.4
sqlparse==0.5.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
mysqlclient==2.2.4
